Replaced old GUI and updated README
